split token from implementation
