+++ src/packages/documents/document-types/repository/detail/document-type-detail.store.ts
@@ -1,7 +1,7 @@
 import type { UmbDocumentTypeDetailModel } from '../../types.js';
-import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 import { UmbDetailStoreBase } from '@umbraco-cms/backoffice/store';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
+import { UMB_DOCUMENT_TYPE_DETAIL_STORE_CONTEXT } from './document-type-detail.store.context-token.js';
 
 /**
  * @export
@@ -19,9 +19,5 @@ export class UmbDocumentTypeDetailStore extends UmbDetailStoreBase<UmbDocumentTy
 		super(host, UMB_DOCUMENT_TYPE_DETAIL_STORE_CONTEXT.toString());
 	}
 }
-
-export const UMB_DOCUMENT_TYPE_DETAIL_STORE_CONTEXT = new UmbContextToken<UmbDocumentTypeDetailStore>(
-	'UmbDocumentTypeDetailStore',
-);
 
 export { UmbDocumentTypeDetailStore as api };
