+++ src/packages/documents/document-types/repository/detail/index.ts
@@ -1,3 +1,3 @@
 export { UmbDocumentTypeDetailRepository } from './document-type-detail.repository.js';
 export { UMB_DOCUMENT_TYPE_DETAIL_REPOSITORY_ALIAS, UMB_DOCUMENT_TYPE_DETAIL_STORE_ALIAS } from './manifests.js';
-export { UMB_DOCUMENT_TYPE_DETAIL_STORE_CONTEXT } from './document-type-detail.store.js';
+export { UMB_DOCUMENT_TYPE_DETAIL_STORE_CONTEXT } from './document-type-detail.store.context-token.js';
